added more type-parameters and removed unused casts

+++ prism/src/explicit/DTMCSimple.java
@@ -312,8 +312,8 @@ public class DTMCSimple extends DTMCExplicit implements ModelSimple
 		distr = transitionList.get(s);
 		d = 0.0;
 		for (Map.Entry<Integer, Double> e : distr) {
-			k = (Integer) e.getKey();
-			prob = (Double) e.getValue();
+			k = e.getKey();
+			prob = e.getValue();
 			d += prob * vect[k];
 		}
 
@@ -331,8 +331,8 @@ public class DTMCSimple extends DTMCExplicit implements ModelSimple
 		diag = 1.0;
 		d = 0.0;
 		for (Map.Entry<Integer, Double> e : distr) {
-			k = (Integer) e.getKey();
-			prob = (Double) e.getValue();
+			k = e.getKey();
+			prob = e.getValue();
 			if (k != s) {
 				d += prob * vect[k];
 			} else {
@@ -355,8 +355,8 @@ public class DTMCSimple extends DTMCExplicit implements ModelSimple
 		distr = transitionList.get(s);
 		d = mcRewards.getStateReward(s);
 		for (Map.Entry<Integer, Double> e : distr) {
-			k = (Integer) e.getKey();
-			prob = (Double) e.getValue();
+			k = e.getKey();
+			prob = e.getValue();
 			d += prob * vect[k];
 		}
 
@@ -378,8 +378,8 @@ public class DTMCSimple extends DTMCExplicit implements ModelSimple
 		for (i = 0; i < numStates; i++) {
 			distr = transitionList.get(i);
 			for (Map.Entry<Integer, Double> e : distr) {
-				j = (Integer) e.getKey();
-				prob = (Double) e.getValue();
+				j = e.getKey();
+				prob = e.getValue();
 				result[j] += prob * vect[i];
 			}
 		}
